requirements.txt
paho-mqtt
tinytuya
